refactor: render main layout
